+++ src/registries/ecr/repos.ts
@@ -5,7 +5,6 @@ export const getRepos = async (context: any, host: string, namespace: string): P
   const records: any[] = []
   try {
     const resp = await context.ecr.describeRepositories().promise()
-    console.log(resp)
 
     resp.repositories.forEach((repo: any) => {
       records.push({
@@ -13,8 +12,10 @@ export const getRepos = async (context: any, host: string, namespace: string): P
         host,
         namespace,
         repo: repo.repositoryName,
-        isPrivate: true,
+        arn: repo.repositoryArn,
         created: repo.createdAt,
+        imageImmutability: repo.imageTagMutability,
+        scanOnPush: typeof (repo.imageScanningConfiguration) === 'undefined' ? false : repo.imageScanningConfiguration.scanOnPush,
       })
     })