+++ src/registries/gcr/images.ts
@@ -31,13 +31,14 @@ export const getImages = async (
     const records: any[] = []
     Object.keys(resp.data.manifest).forEach((digest) => {
       const manifest = resp.data.manifest[digest]
+      console.log(manifest)
       records.push({
         registry: context.name,
         host,
         namespace,
         repo,
         digest,
-        tag: manifest.tag.length === 0 ? '' : manifest.tag[0],
+        tags: manifest.tag,
         size: manifest.imageSizeBytes,
         created: new Date(parseInt(manifest.timeCreatedMs, 10)).toISOString(),
         pushed: new Date(parseInt(manifest.timeUploadedMs, 10)).toISOString(),